pyproject.toml
[project]
name = "machine-learning-zoomcamp-2025-capstone-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cairosvg>=2.8.2",
    "fastapi>=0.128.0",
    "ipykernel>=7.1.0",
    "kaggle>=1.8.3",
    "onnx>=1.20.1",
    "onnxruntime>=1.23.2",
    "onnxscript>=0.5.7",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "python-chess>=1.999",
    "seaborn>=0.13.2",
    "torch>=2.9.1",
    "uvicorn>=0.40.0",
]
